Splitted GwtDeviceService into GwtDeviceService and GwtDeviceManagementService
