+++ org.eclipse.kapua.app.console/src/main/java/org/eclipse/kapua/app/console/client/widget/dialog/SyncPackageUninstallDialog.java
@@ -15,17 +15,17 @@ package org.eclipse.kapua.app.console.client.widget.dialog;
 import org.eclipse.kapua.app.console.client.util.DialogUtils;
 import org.eclipse.kapua.app.console.shared.model.GwtDeploymentPackage;
 import org.eclipse.kapua.app.console.shared.model.GwtDevice;
-import org.eclipse.kapua.app.console.shared.service.GwtDeviceService;
-import org.eclipse.kapua.app.console.shared.service.GwtDeviceServiceAsync;
+import org.eclipse.kapua.app.console.shared.service.GwtDeviceManagementService;
+import org.eclipse.kapua.app.console.shared.service.GwtDeviceManagementServiceAsync;
 
 import com.google.gwt.core.client.GWT;
 import com.google.gwt.user.client.rpc.AsyncCallback;
 
 public class SyncPackageUninstallDialog extends PackageUninstallDialog
 {
-    private final GwtDeviceServiceAsync gwtDeviceService = GWT.create(GwtDeviceService.class);
+    private final GwtDeviceManagementServiceAsync gwtDeviceManagementService = GWT.create(GwtDeviceManagementService.class);
 
-    private GwtDevice                   m_selectedDevice;
+    private GwtDevice m_selectedDevice;
 
     public SyncPackageUninstallDialog(GwtDevice selectedDevice,
                                       GwtDeploymentPackage selectedDeploymentPackage)
@@ -46,26 +46,26 @@ public class SyncPackageUninstallDialog extends PackageUninstallDialog
     @Override
     public void submit()
     {
-        gwtDeviceService.uninstallDeploymentPackage(m_token,
-                                                    m_selectedDevice,
-                                                    m_selectedDeploymentPackage.getUnescapedName(),
-                                                    new AsyncCallback<Void>() {
-                                                        @Override
-                                                        public void onSuccess(Void arg0)
-                                                        {
-                                                            m_exitStatus = true;
-                                                            m_exitMessage = MSGS.packageUninstallSyncSuccess();
-                                                            hide();
-                                                        }
+        gwtDeviceManagementService.uninstallDeploymentPackage(m_token,
+                                                              m_selectedDevice,
+                                                              m_selectedDeploymentPackage.getUnescapedName(),
+                                                              new AsyncCallback<Void>() {
+                                                                  @Override
+                                                                  public void onSuccess(Void arg0)
+                                                                  {
+                                                                      m_exitStatus = true;
+                                                                      m_exitMessage = MSGS.packageUninstallSyncSuccess();
+                                                                      hide();
+                                                                  }
 
-                                                        @Override
-                                                        public void onFailure(Throwable caught)
-                                                        {
-                                                            m_exitStatus = false;
-                                                            m_exitMessage = MSGS.packageInstallSyncFailure(caught.getMessage());
-                                                            hide();
-                                                        }
-                                                    });
+                                                                  @Override
+                                                                  public void onFailure(Throwable caught)
+                                                                  {
+                                                                      m_exitStatus = false;
+                                                                      m_exitMessage = MSGS.packageInstallSyncFailure(caught.getMessage());
+                                                                      hide();
+                                                                  }
+                                                              });
     }
 
     @Override
